Mar_RAg/requirements.txt
streamlit==1.41.1
Werkzeug==3.1.3
python-dotenv==1.0.1
langchain==0.2.7
langchain-community==0.2.7
langchain-openai==0.1.15
openai==1.58.1
loguru == 0.7.2
unstructured==0.16.11